be-django/requirements.txt
django~=5.1.0
asgiref
django-environ
django-debug-toolbar

django-lifecycle~=1.2.4

django-cors-headers
django-crispy-forms
crispy-tailwind

django-pyhub-ai>=0.6.2

psycopg2-binary
gunicorn
uvicorn[standard]
whitenoise
